Simplify code
