+++ WalletWasabi.Fluent/ViewModels/CoinControl/DataGridSource.cs
@@ -1,0 +1,147 @@
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using Avalonia.Controls;
+using Avalonia.Controls.Models.TreeDataGrid;
+using Avalonia.Controls.Templates;
+using NBitcoin;
+using WalletWasabi.Fluent.Helpers;
+using WalletWasabi.Fluent.TreeDataGrid;
+using WalletWasabi.Fluent.ViewModels.CoinControl.Core;
+using WalletWasabi.Fluent.Views.CoinControl.Core.Cells;
+using WalletWasabi.Fluent.Views.CoinControl.Core.Headers;
+
+namespace WalletWasabi.Fluent.ViewModels.CoinControl;
+
+public static class DataGridSource
+{
+	private static Comparison<TSource?> SortAscending<TSource, TProperty>(Func<TSource, TProperty> selector)
+	{
+		return (x, y) => Comparer<TProperty>.Default.Compare(selector(x!), selector(y!));
+	}
+
+	private static Comparison<TSource?> SortDescending<TSource, TProperty>(Func<TSource, TProperty> selector)
+	{
+		return (x, y) => Comparer<TProperty>.Default.Compare(selector(y!), selector(x!));
+	}
+
+	private static int GetLabelPriority(CoinControlItemViewModelBase coin)
+	{
+		if (coin.Labels == CoinPocketHelper.PrivateFundsText)
+		{
+			return 3;
+		}
+
+		if (coin.Labels == CoinPocketHelper.SemiPrivateFundsText)
+		{
+			return 2;
+		}
+
+		return 1;
+	}
+
+	private static int GetIndicatorPriority(CoinControlItemViewModelBase x)
+	{
+		if (x.IsCoinjoining)
+		{
+			return 1;
+		}
+
+		if (x.BannedUntilUtc.HasValue)
+		{
+			return 2;
+		}
+
+		if (!x.IsConfirmed)
+		{
+			return 3;
+		}
+
+		return 0;
+	}
+	
+	private static IColumn<CoinControlItemViewModelBase> ChildrenColumn()
+	{
+		return new HierarchicalExpanderColumn<CoinControlItemViewModelBase>(
+			SelectionColumn(),
+			group => group.Children,
+			node => node.Children.Count > 1,
+			node => node.IsExpanded);
+	}
+
+	private static TemplateColumn<CoinControlItemViewModelBase> SelectionColumn()
+	{
+		return new TemplateColumn<CoinControlItemViewModelBase>(
+			"",
+			new FuncDataTemplate<CoinControlItemViewModelBase>(
+				(_, _) => new SelectionCellView(),
+				true),
+			GridLength.Auto);
+	}
+
+	private static IColumn<CoinControlItemViewModelBase> AmountColumn()
+	{
+		return new PlainTextColumn<CoinControlItemViewModelBase>(
+			"Amount",
+			node => node.Amount.ToFormattedString(),
+			GridLength.Auto,
+			new ColumnOptions<CoinControlItemViewModelBase>
+			{
+				CompareAscending = SortAscending<CoinControlItemViewModelBase, Money>(x => x.Amount),
+				CompareDescending = SortDescending<CoinControlItemViewModelBase, Money>(x => x.Amount)
+			});
+	}
+
+	private static IColumn<CoinControlItemViewModelBase> IndicatorsColumn()
+	{
+		return new TemplateColumn<CoinControlItemViewModelBase>(
+			"",
+			new FuncDataTemplate<CoinControlItemViewModelBase>((_, _) => new IndicatorsCellView(), true),
+			GridLength.Auto,
+			new ColumnOptions<CoinControlItemViewModelBase>
+			{
+				CompareAscending = SortAscending<CoinControlItemViewModelBase, int>(GetIndicatorPriority),
+				CompareDescending = SortDescending<CoinControlItemViewModelBase, int>(GetIndicatorPriority)
+			});
+	}
+
+	private static IColumn<CoinControlItemViewModelBase> AnonymityScoreColumn()
+	{
+		return new PlainTextColumn<CoinControlItemViewModelBase>(
+			new AnonymityScoreHeaderView(),
+			node => node.AnonymityScore.ToString(),
+			GridLength.Auto,
+			new TextColumnOptions<CoinControlItemViewModelBase>
+			{
+				CompareAscending = SortAscending<CoinControlItemViewModelBase, int?>(b => b.AnonymityScore),
+				CompareDescending = SortDescending<CoinControlItemViewModelBase, int?>(b => b.AnonymityScore)
+			});
+	}
+
+	private static IColumn<CoinControlItemViewModelBase> PocketColumn()
+	{
+		return new TemplateColumn<CoinControlItemViewModelBase>(
+			"Pocket",
+			new FuncDataTemplate<CoinControlItemViewModelBase>((_, _) => new LabelsCellView(), true),
+			GridLength.Star,
+			new ColumnOptions<CoinControlItemViewModelBase>
+			{
+				CompareAscending = SortAscending<CoinControlItemViewModelBase, int>(GetLabelPriority),
+				CompareDescending = SortDescending<CoinControlItemViewModelBase, int>(GetLabelPriority)
+			});
+	}
+
+	public static HierarchicalTreeDataGridSource<CoinControlItemViewModelBase> CreateSource(IEnumerable<CoinControlItemViewModelBase> source)
+	{
+		return new HierarchicalTreeDataGridSource<CoinControlItemViewModelBase>(source)
+		{
+			Columns =
+			{
+				ChildrenColumn(),
+				IndicatorsColumn(),
+				AmountColumn(),
+				AnonymityScoreColumn(),
+				PocketColumn()
+			}
+		};
+	}
+}
